refactoring backend & fixing websocket connection & avatar on chat
