+++ data/backend/src/app.module.ts
@@ -6,12 +6,17 @@ import { AppService } from './app.service';
 
 import { UserService } from './services/user.service';
 
-import { ChatModule } from './chatModule/chat.module';
-import { GameModule } from './gameModule/game.module';
 import { User } from './entities/user.entity';
 
-import { ServeStaticModule } from '@nestjs/serve-static';
-import { join } from 'path';
+import { ChatController } from './chatModule/chat.controller';
+import { RoomService } from './services/room.service';
+import { Room } from './entities/room.entity';
+import { Achievements } from './entities/achievements.entity';
+import { MatchHistory } from './entities/matchHistory.entity';
+import { Password } from './entities/password.entity';
+import { Stats } from 'fs';
+import { UsersList } from './entities/usersList.entity';
+import { Gateway } from './gateway/gateway';
 
 @Module({
   imports: [
@@ -27,11 +32,17 @@ import { join } from 'path';
 		synchronize: true,
 		autoLoadEntities: true,
 	}),
-	ChatModule,
-	GameModule,
-	TypeOrmModule.forFeature([User]),
+	TypeOrmModule.forFeature([
+		User,
+		Room,
+		Achievements,
+		MatchHistory,
+		Password,
+		Stats,
+		UsersList
+	]),
   ],
-  controllers: [AppController],
-  providers: [UserService, AppService],
+  controllers: [AppController, ChatController],
+  providers: [UserService, AppService, RoomService, Gateway],
 })
 export class AppModule {}
