+++ data/frontend/src/index.tsx
@@ -5,7 +5,6 @@ import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux'
 import store from './store/default';
 
-
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
